requirements.txt
django==2.2.10
djangorestframework==3.9.1
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.4.0
idna==2.9
PyJWT==1.7.1
python-dateutil==2.8.1
python-magic==0.4.15
requests==2.23.0
urllib3==1.25.9
websocket-client==0.48.0
MarkupSafe==1.1.1
coreapi==2.3.3
coreschema==0.0.4
django-rest-swagger==2.2.0
itypes==1.2.0
jinja2==2.11.2
openapi-codec==1.3.2
simplejson==3.17.0
uritemplate==3.0.1